Adjust position before its sent to the renderers

+++ client/view/ActorRenderer.h
@@ -13,9 +13,9 @@ enum ActorRendererType {
 class ActorRendererr {
 public:
     ActorRendererr(SDL2pp::Renderer *renderer, SDL2pp::Texture *sprites,
-                   float pos_x, float pos_y);
+                   int pos_x, int pos_y);
 
-    void update(float pos_x, float pos_y, int dir_x, int dir_y);
+    void update(int pos_x, int pos_y, int dir_x, int dir_y);
 
     virtual void render() = 0;
 
@@ -24,8 +24,8 @@ public:
 protected:
     SDL2pp::Renderer *renderer;
     SDL2pp::Texture *sprites;
-    float pos_x;
-    float pos_y;
+    int pos_x;
+    int pos_y;
     int dir_x;
     int dir_y;
 };
@@ -45,8 +45,8 @@ public:
 class ActorRendererFactory {
 public:
     explicit ActorRendererFactory(SDL2pp::Renderer * renderer);
-    ActorRendererr* build_actor_renderer(std::string tile_type, float pos_x,
-                                       float pos_y);
+    ActorRendererr* build_actor_renderer(std::string tile_type, int pos_x,
+                                         int pos_y);
     virtual ~ActorRendererFactory() {}
 private:
     SDL2pp::Renderer *renderer;
